# perf(ext/streams): optimize async iterator